+++ miniconf/src/node.rs
@@ -130,7 +130,7 @@ impl Transcode for () {
 /// The path will either be empty or start with the separator.
 ///
 /// * `path: T`: A `Write` to write the separators and node names into during `Transcode`.
-///   See also [TreeKey::metadata()] and `Metadata::max_length()` for upper bounds
+///   See also [TreeKey::walk()] and `Metadata::max_length()` for upper bounds
 ///   on path length. Can also be a `AsRef<str>` to implement `IntoKeys` (see [`KeysIter`]).
 /// * `const S: char`: The path hierarchy separator to be inserted before each name,
 ///   e.g. `'/'`.